Fixed Memory Leak and Removed Outdated BreakPointHandler
One of the useEffect hooks in useBreakpoints caused a memory
leak. There was also a leak in the header component. I refactored
useBreakpoints to no longer rely on the BreakPointHandler and
instead rewrote the methods in the useBreakpoints for a Single
Source of Truth. This fix drastically improves the quality
of the site and includes several minor fixes to typos
that caused an issue with the styling.